requirements.txt
supabase
python-dotenv
fastapi
uvicorn
python-dateutil
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
deprecated==1.2.14
google-adk==1.0.0
mcp==1.9.1
python-jose
mcp-youtube-searc